Refactor: Extract chronicle state helpers and version footer rendering
